Add thox amongst the domains
thox is an OS project for ComputerCraft. For now it mainly consists of a
documentation of external components and of the OS design, not a valid
implementation yet; but as it documents many components from other OSes
and projects, such as OPUS and OneOS, I consider it a valid proposal to
be submitted for the madefor.cc domain.